+++ node-crawler-to-dynamoDB-typescript/.env.stage
@@ -1,0 +1,9 @@
+NODE_ENV=stage
+# TODO DYNAMODB 테이블 명을 입력해 주세요
+DYNAMODB_TABLE_NAME=
+# TODO 데이터를 가져올 RESOURCS API URL
+BASE_URL=
+# TODO BASE_URL 의 TOKEN 값
+TOKEN=
+# TODO AWS 리전 값 디폴트는 'ap-southeast-1' 싱가폴
+REGION=
