+++ node-crawler-to-dynamoDB-typescript/.npmignore
