+++ node-crawler-to-dynamoDB-typescript/.prettierrc
