+++ node-crawler-to-dynamoDB-typescript/src/getData.ts
