+++ node-crawler-to-dynamoDB-typescript/src/types.ts
