+++ node-crawler-to-dynamoDB-typescript/tsconfig.json
