--- node-crawler-typescript/.env.dev
@@ -1,7 +1,0 @@
-NODE_ENV=dev
-DYNAMODB_TABLE_NAME= ## TODO DYNAMODB 테이블 명을 입력해 주세요
-BASE_URL= ## TODO 데이터를 가져올 RESOURCS API URL
-## TODO BASE_URL 의 TOKEN 값
-TOKEN=
-## TODO AWS 리전 값 디폴트는 'ap-southeast-1' 싱가폴
-REGION=
